pyproject.toml
[project]
name = "aisdecoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Giampaolo Cimino", email = "gcimino@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "libais>=0.17",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
]

[project.scripts]
aisdecoder = "aisdecoder.cli.main:main" 

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "assertpy>=1.1",
    "deptry>=0.22.0",
    "lxml>=5.3.1",
    "mypy>=1.14.1",
    "pandas>=2.2.3",
    "xlsx2csv>=0.8.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aisdecoder"]
